Hall Service updated

+++ venueManagement/src/main/java/et/tk/api/venueManagement/hall/HallService.java
@@ -1,5 +1,6 @@
 package et.tk.api.venueManagement.hall;
 
+import et.tk.api.venueManagement.seat.Seat;
 import et.tk.api.venueManagement.seat.SeatRepository;
 import et.tk.api.venueManagement.venue.VenueRepository;
 import lombok.AllArgsConstructor;
@@ -7,7 +8,9 @@ import org.springframework.cglib.core.CollectionUtils;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 @AllArgsConstructor
@@ -17,41 +20,70 @@ public class HallService {
     private HallRepository hallRepository;
     private VenueRepository venueRepository;
 
-    public HallDto getHall(String hallId) {
-        Optional<Hall> hallOptional = hallRepository.findById(hallId);
+    public String createHall(String venueId, Hall hall) {
+        hall.setId(null);
+        hall.setName(hall.getName().toLowerCase());
+        hall.setVenueId(venueId);
+        if (venueRepository.findById(venueId).isEmpty())
+            return "venue";
+        List<Hall> nameCheck = hallRepository.findAll();
+        CollectionUtils.filter(nameCheck, o -> ((Hall) o).getVenueId().equals(venueId));
+        CollectionUtils.filter(nameCheck, o -> ((Hall) o).getName().equals(hall.getName()));
 
-        return hallOptional.map(HallDto::new).orElse(null);
+        if (nameCheck.isEmpty()) {
+            hallRepository.save(hall);
+            return "created";
+        } else
+            return "name";
     }
 
-    public String updateHall(String hallId, HallDto hallDto) {
+    public List<Hall> getHallsByVenueId(String id) {
+        List<Hall> halls = hallRepository.findAll();
+        CollectionUtils.filter(halls, o -> ((Hall) o).getVenueId().equals(id));
+        if (halls.isEmpty())
+            return null;
+        return halls;
+    }
+
+    public Hall getHallById(String id) {
+        return hallRepository.findById(id).orElse(null);
+    }
+
+    public String updateHall(String hallId, Hall hall) {
         Optional<Hall> hallOptional = hallRepository.findById(hallId);
         if (hallOptional.isEmpty())
             return "hall";
 
-        Hall hall = hallOptional.get();
+        if (!Objects.equals(hall.getVenueId(), hallOptional.get().getVenueId()))
+            return "venue id";
+
+        Hall backup = hallOptional.get();
+        hall.setId(backup.getId());
 
         hallRepository.deleteById(hallId);
 
-        List<Hall> nameCheck = hallRepository.findByVenueId(hall.getVenueId());
-        CollectionUtils.filter(nameCheck, o -> ((Hall) o).getName().equals(hallDto.getName()));
+        List<Hall> nameCheck = hallRepository.findByVenueId(backup.getVenueId());
+        CollectionUtils.filter(nameCheck, o -> ((Hall) o).getName().equals(hall.getName()));
 
         if (nameCheck.isEmpty()) {
-            Hall updatedHall = new Hall(hallDto);
-            updatedHall.setId(hall.getId());
-            updatedHall.setVenueId(hall.getVenueId());
-            hallRepository.save(updatedHall);
+            hall.setId(backup.getId());
+            hallRepository.save(hall);
             return "updated";
         } else {
-            hallRepository.save(hall);
+            hallRepository.save(backup);
             return "name";
         }
     }
 
-    public String deleteHall(String hallId) {
-        Optional<Hall> hallOptional = hallRepository.findById(hallId);
+    public String deleteHall(String id) {
+        Optional<Hall> hallOptional = hallRepository.findById(id);
 
         if (hallOptional.isPresent()) {
-            hallRepository.deleteById(hallId);
+            List<Seat> seats = seatRepository.findByHallId(id);
+            for (Seat seat:seats){
+                seatRepository.deleteById(seat.getId());
+            }
+            hallRepository.deleteById(id);
             return "deleted";
         } else {
             return "not found";
